requirements.txt
pyyaml==5.4.1
jinja2==2.11.3
pytest==6.2.2